CR FIX move genric parser and scanner classes to fi.fmi.avi.converter.iwxxm.generic
