+++ src/main/java/fi/fmi/avi/converter/iwxxm/generic/GenericAviationWeatherMessageParser.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
+package fi.fmi.avi.converter.iwxxm.generic;
 
 import org.w3c.dom.Document;
