+++ src/main/java/fi/fmi/avi/converter/iwxxm/generic/GenericBulletinIWXXMParser.java
@@ -1,4 +1,4 @@
-package fi.fmi.avi.converter.iwxxm.bulletin.v1_2;
+package fi.fmi.avi.converter.iwxxm.generic;
 
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoField;
@@ -9,6 +9,8 @@ import org.w3c.dom.Document;
 import fi.fmi.avi.converter.ConversionException;
 import fi.fmi.avi.converter.ConversionHints;
 import fi.fmi.avi.converter.iwxxm.bulletin.AbstractBulletinIWXXMParser;
+import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.BulletinProperties;
+import fi.fmi.avi.converter.iwxxm.bulletin.v1_2.MeteorologicalBulletinIWXXMScanner;
 import fi.fmi.avi.model.GenericAviationWeatherMessage;
 import fi.fmi.avi.model.bulletin.BulletinHeading;
 import fi.fmi.avi.model.bulletin.GenericMeteorologicalBulletin;
