+++ src/test/resources/fi/fmi/avi/converter/iwxxm/generic/iwxxm-30-cancellation-bulletin.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<collect:MeteorologicalBulletin xmlns:collect="http://def.wmo.int/collect/2014"
+  xmlns:gml="http://www.opengis.net/gml/3.2"
+  xmlns:aixm="http://www.aixm.aero/schema/5.1.1"
+  xmlns:iwxxm="http://icao.int/iwxxm/3.0"
+  xmlns:xlink="http://www.w3.org/1999/xlink"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  gml:id="bulletin-12345"
+  xsi:schemaLocation="http://icao.int/iwxxm/3.0 http://schemas.wmo.int/iwxxm/3.0/iwxxm.xsd http://def.wmo.int/collect/2014 http://schemas.wmo.int/collect/1.2/collect.xsd">
+  <collect:meteorologicalInformation>
+<iwxxm:TAF
+  gml:id="uuid.dfbfe27c-3478-4efc-9800-124b90030e39"
+  reportStatus="AMENDMENT"
+  permissibleUsage="OPERATIONAL"
+  isCancelReport="true">
+
+  <iwxxm:issueTime>
+    <gml:TimeInstant gml:id="uuid.2a142955-e0ef-4164-a550-ec80a3258150">
+      <gml:timePosition>2012-08-16T15:00:00Z</gml:timePosition>
+    </gml:TimeInstant>
+  </iwxxm:issueTime>
+
+  <iwxxm:aerodrome>
+    <aixm:AirportHeliport gml:id="uuid.65730df9-a2fd-4815-9d3f-d6f6770bc6cf">
+      <aixm:timeSlice>
+        <aixm:AirportHeliportTimeSlice gml:id="uuid.30a6076d-48ec-42df-bcd5-bf85220fe274">
+          <gml:validTime/>
+          <aixm:interpretation>SNAPSHOT</aixm:interpretation>
+          <aixm:designator>YUDO</aixm:designator>
+          <aixm:name>DONLON/INTERNATIONAL</aixm:name>
+          <aixm:locationIndicatorICAO>YUDO</aixm:locationIndicatorICAO>
+        </aixm:AirportHeliportTimeSlice>
+      </aixm:timeSlice>
+    </aixm:AirportHeliport>
+  </iwxxm:aerodrome>
+
+  <iwxxm:cancelledReportValidPeriod>
+    <gml:TimePeriod gml:id="uuid.b171a7e2-7fdf-496c-ada8-6cbc6c606291">
+      <gml:beginPosition>2012-08-16T00:00:00Z</gml:beginPosition>
+      <gml:endPosition>2012-08-16T18:00:00Z</gml:endPosition>
+    </gml:TimePeriod>
+  </iwxxm:cancelledReportValidPeriod>
+
+</iwxxm:TAF>
+  </collect:meteorologicalInformation>
+  <collect:bulletinIdentifier>A_LYXX31YUDO251600_C_YUDO_201208251600--.xml</collect:bulletinIdentifier>
+</collect:MeteorologicalBulletin>
